requirement.txt
wandb
lmdb
pyarrow
regex
ftfy
loguru
pycocotools
matplotlib
tqdm
pickle5
snntorch
torchvision
torchaudio